% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendInTrend.R
\name{OR}
\alias{OR}
\title{An Odds Ratio Estimation Function}
\usage{
OR(n11, n10, n01, n00)
}
\arguments{
\item{n11}{A G by T matrix with n11[i,j] being the counts of positive outcomes among treated subjests within group i at time j;}

\item{n10}{A G by T matrix with n10[i,j] being the counts of negative outcomes among treated subjests within group i at time j;}

\item{n01}{A G by T matrix with n01[i,j] being the counts of positive outcomes among control subjests within group i at time j;}

\item{n00}{A G by T matrix with n00[i,j] being the counts of negative outcomes among control subjests within group i at time j.}
}
\value{
ORs ratio and 95\% conference interval
}
\description{
estimate causal odds ratio given trends in exposure prevalence and outcome frequencies of stratified data.
}
\examples{
\donttest{
data <- GenData()
n11 <- data[[1]]
n10 <- data[[2]]
n01 <- data[[3]]
n00 <- data[[4]]
results <- OR(n11,n10,n01,n00)
}
}

